Make memory protection active on Aranym by default again. First read mint.ini and *then* try to override its values (namely no_mem_prot flag). Check (very simply) if the PMMU is present on CPU, override no_mem_prot if it is not. Make the CT60 PMMU hack actually visible, before no_mem_prot was always overriden by mint.ini.
Contributed by Miro Kropacek.

+++ sys/global.c
@@ -40,7 +40,7 @@ struct global global =
 };
 
 long mcpu = 0;
-long pmmu = 0;
+short pmmu = 0;
 short fpu = 0;
 #ifdef __mcoldfire__
 bool coldfire_68k_emulation = false;
